feat(5.12): use modules
